requirements.txt
pyyaml>=6.0
rich>=13.0.0
openai>=1.0.0
anthropic>=0.3.0
requests>=2.28.0